Plataforma de Saúde Comunitária/backend/requirements.txt
Flask
azure-cosmos
